requirements.txt
numpy
pandas
openai
streamlit
python-dotenv
crewai
tavily-python
